refactor: update DrawerFilter and routes to use new utils/filter path for better organization and type clarity
