+++ app/lib/utils/filter.ts
@@ -2,7 +2,7 @@ import type {
   ProductFilter,
 } from '@shopify/hydrogen/storefront-api-types';
 import type {Location, useLocation} from '@remix-run/react';
-import { FILTER_URL_PREFIX } from "./utils/const";
+import { FILTER_URL_PREFIX } from "./const";
 
 export type AppliedFilter = {
   label: string;
