+++ app/routes/($locale).collections.$handle.tsx
@@ -14,7 +14,7 @@ import {json, type LoaderFunctionArgs} from '@shopify/remix-oxygen';
 import {routeHeaders} from '~/data/cache';
 import {COLLECTION_QUERY} from '~/data/queries';
 import {FILTER_URL_PREFIX, PAGINATION_SIZE} from '~/lib/utils/const';
-import {SortParam} from '~/lib/filter';
+import {SortParam} from '~/lib/utils/filter';
 import {seoPayload} from '~/lib/seo.server';
 import {parseAsCurrency} from '~/lib/utils';
 import {WeaverseContent} from '~/weaverse';
